requirements.txt
# === AI LEARNING AGENT — Dependencies ===

# FastAPI Backend
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.12
pydantic==2.10.3
pydantic-settings==2.6.1

# Streamlit Frontend
streamlit==1.40.2
streamlit-extras==0.4.7

# RAG & LLM
openai==1.57.0          # OpenRouter is OpenAI-compatible
langchain==0.3.9
langchain-community==0.3.9
langchain-openai==0.2.11

# Vector Store
chromadb==0.5.20

# Embeddings (free, runs locally)
sentence-transformers==3.3.1

# Database (SQLite for dev, swap to PostgreSQL later)
sqlalchemy==2.0.36
aiosqlite==0.20.0

# Auth & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1

# Utilities
httpx==0.27.2
pandas==2.2.3
PyPDF2==3.0.1
python-docx==1.1.2
schedule==1.2.2
